requirements.txt
flask==3.0.2
gunicorn==21.2.0
tensorflow-cpu==2.15.0
tf-keras==2.15.0
numpy
pillow
opencv-python-headless
moviepy